requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
callback==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
dnspython==2.3.0
fastapi==0.95.2
frozenlist==1.3.3
functions==0.7.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
langchain==0.0.171
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
schemas==0.7.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.27.0
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
